Late commit- Bug fix -onset null pointer & playingstate crashing
Under the setupcards, currentdialogueIndex randomizer was setup to choose from a fixed range. Changed the algorithm to dynamically adjust with the current's card questions' size.
Upon performing change card, a crash might happen as the loop continuously runs when all of the cards max number of times has reached their limit, and no other cards was left to be chosen. Removed maxnumber of times algorithm. An update may be released soon to handle randomizer better.